Correção no tipo do cash

+++ core/models.py
@@ -390,8 +390,8 @@ class Professional(models.Model):
 
     @property
     def cash(self):
-        cash_in = int(self.receipts.all().aggregate(models.Sum('value'))['value__sum'] or 0)
-        cash_out = int(self.cash_outs.all().aggregate(models.Sum('value'))['value__sum'] or 0)
+        cash_in = float(self.receipts.all().aggregate(models.Sum('value'))['value__sum'] or 0)
+        cash_out = float(self.cash_outs.all().aggregate(models.Sum('value'))['value__sum'] or 0)
         return cash_in - cash_out
 
     def create_recipient(self, agency, agency_dv, bank_code, account, account_dv, legal_name, account_type):
